Switch to the ldap.DialURL function and introduce an option for TLS verification
Currently we're using ldap.Dial which offers only the option to use plain
LDAP. Using the new DialURL offers a way to pass full URIs, with the
library taking care using TLS or not depending on the schema provided.

This commit also introduced an option to disable TLS certificate
verification when LDAPS is being used.