Retrieve client information from spoe message
The client information are statically defined in agent config file.
Goal of this patch is to be able to send those information
directly via the spoe messages, and therefore be able to configure
them directly on HAProxy config file.

For the test/docker, since we can run only type of environment, run the
one with dynamic client info option enabled.