Migrate library haproxy-spoe-go
Since our library https://github.com/criteo/haproxy-spoe-go has not
been updated since a while and that some part of the specification
are not well respected (here field engine-id and capability
pipelining should be optional), migrate to a better maintained
library.

Using this library, we could now enable the spop-check in HAProxy
for this component.